Add option to set the times header
If the column is empty, each EAN code is generated only once.
If it is not empty, the column must contain only numbers,
and each EAN is generated N times based on the number in that column.
EAN codes are processed line by line, so identical EANs appear next to each other.